moved some test cases
